requirements.txt
colorama
python-gitlab
python-dateutil
humanize
